# More efficient default encoding to strings