Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                                  (size)   Mode  Cnt         Score         Error  Units
ADTWriting.zioJson                            N/A  thrpt    5   2446477.121 ±    3828.169  ops/s
AnyValsWriting.zioJson                        N/A  thrpt    5   4285913.815 ±   78292.399  ops/s
ArrayBufferOfBooleansWriting.zioJson          128  thrpt    5   1048903.155 ±   32001.706  ops/s
ArrayOfBigDecimalsWriting.zioJson             128  thrpt    5     86577.324 ±    3435.528  ops/s
ArrayOfBigIntsWriting.zioJson                 128  thrpt    5     96500.939 ±    1358.784  ops/s
ArrayOfBooleansWriting.zioJson                128  thrpt    5    954080.727 ±   34228.978  ops/s
ArrayOfBytesWriting.zioJson                   128  thrpt    5   1334865.579 ±   15756.476  ops/s
ArrayOfCharsWriting.zioJson                   128  thrpt    5   1202856.554 ±   34615.643  ops/s
ArrayOfDoublesWriting.zioJson                 128  thrpt    5    188174.779 ±    1474.490  ops/s
ArrayOfDurationsWriting.zioJson               128  thrpt    5    171971.039 ±    3536.193  ops/s
ArrayOfEnumADTsWriting.zioJson                128  thrpt    5    553561.970 ±    9631.389  ops/s
ArrayOfEnumsWriting.zioJson                   128  thrpt    5    671131.338 ±   12585.420  ops/s
ArrayOfFloatsWriting.zioJson                  128  thrpt    5    266833.716 ±    8676.455  ops/s
ArrayOfInstantsWriting.zioJson                128  thrpt    5    164825.490 ±    1095.352  ops/s
ArrayOfIntsWriting.zioJson                    128  thrpt    5    717845.231 ±    9501.502  ops/s
ArrayOfLocalDateTimesWriting.zioJson          128  thrpt    5    232596.453 ±    2859.296  ops/s
ArrayOfLocalDatesWriting.zioJson              128  thrpt    5    378415.567 ±    2655.644  ops/s
ArrayOfLocalTimesWriting.zioJson              128  thrpt    5    321630.004 ±    3527.985  ops/s
ArrayOfLongsWriting.zioJson                   128  thrpt    5    394422.779 ±    7256.495  ops/s
ArrayOfMonthDaysWriting.zioJson               128  thrpt    5    660578.112 ±   24556.711  ops/s
ArrayOfOffsetDateTimesWriting.zioJson         128  thrpt    5    159983.513 ±    1550.934  ops/s
ArrayOfOffsetTimesWriting.zioJson             128  thrpt    5    249814.823 ±    6870.869  ops/s
ArrayOfPeriodsWriting.zioJson                 128  thrpt    5    205605.870 ±    7026.158  ops/s
ArrayOfShortsWriting.zioJson                  128  thrpt    5    910325.023 ±   25976.309  ops/s
ArrayOfUUIDsWriting.zioJson                   128  thrpt    5    171786.704 ±    2487.755  ops/s
ArrayOfYearMonthsWriting.zioJson              128  thrpt    5    565643.079 ±   18377.583  ops/s
ArrayOfYearsWriting.zioJson                   128  thrpt    5    938795.881 ±   14207.508  ops/s
ArrayOfZoneIdsWriting.zioJson                 128  thrpt    5    560009.700 ±    7515.634  ops/s
ArrayOfZoneOffsetsWriting.zioJson             128  thrpt    5    621881.443 ±    9857.014  ops/s
ArrayOfZonedDateTimesWriting.zioJson          128  thrpt    5    137084.167 ±    1834.363  ops/s
ArraySeqOfBooleansWriting.zioJson             128  thrpt    5    979283.444 ±   19265.165  ops/s
Base64Writing.zioJson                         128  thrpt    5   6570745.371 ±   65448.657  ops/s
BigDecimalWriting.zioJson                     128  thrpt    5   1094923.105 ±   20745.723  ops/s
BigIntWriting.zioJson                         128  thrpt    5   1102699.927 ±   15835.377  ops/s
GeoJSONWriting.zioJson                        N/A  thrpt    5     15777.406 ±     240.390  ops/s
GitHubActionsAPIWriting.zioJson               N/A  thrpt    5    604002.506 ±    7590.858  ops/s
GoogleMapsAPIWriting.zioJson                  N/A  thrpt    5     35982.755 ±     346.175  ops/s
IntWriting.zioJson                            N/A  thrpt    5  37856674.221 ± 1067274.716  ops/s
ListOfBooleansWriting.zioJson                 128  thrpt    5   1185395.642 ±   32794.682  ops/s
MapOfIntsToBooleansWriting.zioJson            128  thrpt    5    257214.860 ±    7204.003  ops/s
MutableMapOfIntsToBooleansWriting.zioJson     128  thrpt    5    246645.651 ±    6746.935  ops/s
MutableSetOfIntsWriting.zioJson               128  thrpt    5    537335.499 ±    8506.192  ops/s
NestedStructsWriting.zioJson                  128  thrpt    5    419511.531 ±    3701.376  ops/s
PrimitivesWriting.zioJson                     N/A  thrpt    5   6242401.376 ±  190914.026  ops/s
SetOfIntsWriting.zioJson                      128  thrpt    5    563397.372 ±   18761.546  ops/s
StringOfAsciiCharsWriting.zioJson             128  thrpt    5   8235744.007 ±  178053.061  ops/s
StringOfNonAsciiCharsWriting.zioJson          128  thrpt    5   3203330.710 ±   49931.776  ops/s
VectorOfBooleansWriting.zioJson               128  thrpt    5   1178248.051 ±   20358.365  ops/s
```

#### After
```txt
Benchmark                                  (size)   Mode  Cnt         Score         Error  Units
ADTWriting.zioJson                            N/A  thrpt    5   2491707.239 ±   44895.386  ops/s
AnyValsWriting.zioJson                        N/A  thrpt    5   4311092.014 ±  100336.209  ops/s
ArrayBufferOfBooleansWriting.zioJson          128  thrpt    5    996157.885 ±    6463.267  ops/s
ArrayOfBigDecimalsWriting.zioJson             128  thrpt    5     97507.215 ±    5617.251  ops/s
ArrayOfBigIntsWriting.zioJson                 128  thrpt    5    107400.220 ±    9246.735  ops/s
ArrayOfBooleansWriting.zioJson                128  thrpt    5   1234944.795 ±   54006.461  ops/s
ArrayOfBytesWriting.zioJson                   128  thrpt    5   1277836.743 ±   20299.217  ops/s
ArrayOfCharsWriting.zioJson                   128  thrpt    5   1128524.776 ±   29576.307  ops/s
ArrayOfDoublesWriting.zioJson                 128  thrpt    5    242208.039 ±   10367.750  ops/s
ArrayOfDurationsWriting.zioJson               128  thrpt    5    216276.979 ±    2569.403  ops/s
ArrayOfEnumADTsWriting.zioJson                128  thrpt    5    781166.582 ±   17842.093  ops/s
ArrayOfEnumsWriting.zioJson                   128  thrpt    5    782548.099 ±    8104.795  ops/s
ArrayOfFloatsWriting.zioJson                  128  thrpt    5    292755.364 ±    5647.454  ops/s
ArrayOfInstantsWriting.zioJson                128  thrpt    5    221307.781 ±    5681.666  ops/s
ArrayOfIntsWriting.zioJson                    128  thrpt    5    680636.325 ±    8380.584  ops/s
ArrayOfLocalDateTimesWriting.zioJson          128  thrpt    5    209568.218 ±    1054.299  ops/s
ArrayOfLocalDatesWriting.zioJson              128  thrpt    5    441581.956 ±    4408.013  ops/s
ArrayOfLocalTimesWriting.zioJson              128  thrpt    5    320613.110 ±     880.005  ops/s
ArrayOfLongsWriting.zioJson                   128  thrpt    5    498552.082 ±    8344.910  ops/s
ArrayOfMonthDaysWriting.zioJson               128  thrpt    5    537401.955 ±    4020.940  ops/s
ArrayOfOffsetDateTimesWriting.zioJson         128  thrpt    5    258975.996 ±    1513.878  ops/s
ArrayOfOffsetTimesWriting.zioJson             128  thrpt    5    225722.085 ±    6814.133  ops/s
ArrayOfPeriodsWriting.zioJson                 128  thrpt    5    245310.530 ±    3633.747  ops/s
ArrayOfShortsWriting.zioJson                  128  thrpt    5    895010.540 ±    4207.154  ops/s
ArrayOfUUIDsWriting.zioJson                   128  thrpt    5    220832.419 ±     806.009  ops/s
ArrayOfYearMonthsWriting.zioJson              128  thrpt    5    503854.941 ±    2277.717  ops/s
ArrayOfYearsWriting.zioJson                   128  thrpt    5    820728.520 ±    5668.814  ops/s
ArrayOfZoneIdsWriting.zioJson                 128  thrpt    5    576730.719 ±   19851.234  ops/s
ArrayOfZoneOffsetsWriting.zioJson             128  thrpt    5    571654.905 ±    3420.433  ops/s
ArrayOfZonedDateTimesWriting.zioJson          128  thrpt    5    133038.612 ±     687.499  ops/s
ArraySeqOfBooleansWriting.zioJson             128  thrpt    5   1471592.783 ±   43680.998  ops/s
Base64Writing.zioJson                         128  thrpt    5  12972785.339 ±  358726.596  ops/s
BigDecimalWriting.zioJson                     128  thrpt    5   1100075.426 ±   16711.260  ops/s
BigIntWriting.zioJson                         128  thrpt    5   1117156.257 ±   28000.544  ops/s
GeoJSONWriting.zioJson                        N/A  thrpt    5     16979.473 ±     378.278  ops/s
GitHubActionsAPIWriting.zioJson               N/A  thrpt    5    855308.376 ±   11061.669  ops/s
GoogleMapsAPIWriting.zioJson                  N/A  thrpt    5     42801.236 ±     276.868  ops/s
IntWriting.zioJson                            N/A  thrpt    5  45233221.961 ± 1805425.617  ops/s
ListOfBooleansWriting.zioJson                 128  thrpt    5   1019140.834 ±   10075.400  ops/s
MapOfIntsToBooleansWriting.zioJson            128  thrpt    5    303441.399 ±    7537.146  ops/s
MutableMapOfIntsToBooleansWriting.zioJson     128  thrpt    5    310776.329 ±    2659.552  ops/s
MutableSetOfIntsWriting.zioJson               128  thrpt    5    676033.780 ±   24389.161  ops/s
NestedStructsWriting.zioJson                  128  thrpt    5    244730.670 ±    2176.845  ops/s
PrimitivesWriting.zioJson                     N/A  thrpt    5   5834958.116 ±  451931.165  ops/s
SetOfIntsWriting.zioJson                      128  thrpt    5    718218.891 ±    5242.705  ops/s
StringOfAsciiCharsWriting.zioJson             128  thrpt    5  11733973.120 ±   47722.106  ops/s
StringOfNonAsciiCharsWriting.zioJson          128  thrpt    5   3670975.627 ±   16357.948  ops/s
VectorOfBooleansWriting.zioJson               128  thrpt    5   1145200.410 ±    4321.598  ops/s
```